refactor(cli): extract App layout policy
